finished part4

+++ blog_list/controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('userId', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -17,27 +19,43 @@ blogRouter.get('/:id', async (request, response) => {
 
 blogRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
 
+  if(!user){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    userId: user.id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog.id)
+  await user.save()
   response.status(201).json(savedBlog)
 })
 
 
 blogRouter.delete('/:id', async (request, response) => {
+  console.log(request.user)
+  const user = request.user
+  const blogToDelete = await Blog.findById(request.params.id)
+  if(!user || blogToDelete.userId != user.id){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
 
 blogRouter.put('/:id', async (request, response) => {
   const body = request.body
-
+  const user = request.user
+  if(!user){
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
   const blog = {
     title: body.title,
     author: body.author,
